\name{env_var}
\alias{env_var}
\docType{data}
\title{
Environmental variables for 241 individuals of A.thaliana
}
\description{
A matrix of 241 rows and 18 columns. Global climate and weather data corresponding to individual geographic coordinates were downloaded from the World-Clim  database  (https://worldclim.org).   Eighteen  bioclimatic  variables,  derived from the monthly temperature and rainfall values,  were considered as representing the current environmental matrix.
}
\usage{data("env_var")}

\keyword{datasets}
