\name{genotype_athaliana}
\alias{genotype_athaliana}
\docType{data}
\title{
Genotype from Arabidopsis Thaliana
}
\description{

A matrix of 241 rows and 10 000 columns. It corresponds to 241 swedish plant accessions from The 1,001 Genomes database for Arabidopsis thaliana. The 10,000 filtered SNPs were randomly selected from a large set of variants. The matrix contains no missing genotypes.
}
\usage{data("genotype_athaliana")}


\keyword{datasets}
