\name{pop_athaliana}
\alias{pop_athaliana}
\docType{data}
\title{
Population of individuals from genotype_athaliana
}
\description{
A vector of length 241. The i-th element corresponds to the population of the i-th individuals in dataset genotype_athaliana and env_var. Rhe  individuals  were  clustered  in  groups  based  on  analysis  of  population structure accounting for geographic proximity.
}

\keyword{datasets}
